+++ test/unit/effect/construction.test.js
@@ -23,6 +23,13 @@ describe('Effect construction', function() {
     expect(spy).toHaveBeenCalledWith(repo, true)
   })
 
+  it('errors when given non-POJO options', function() {
+    expect(function() {
+      let repo = new Microcosm()
+      repo.addEffect({}, new Microcosm())
+    }).toThrow(/expected a plain object\. Instead got Microcosm/)
+  })
+
   it('class - with defaults', function() {
     expect.assertions(1)